vision-object-classifier/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
timm>=0.9.0
kaggle>=1.5.0
pytest>=7.0.0
fastapi>=0.104.0
pydantic>=2.5.0
uvicorn>=0.24.0